+++ ESFramework/Core/ESHelpers.h
@@ -221,24 +221,6 @@ NS_INLINE void es_dispatch_sync(dispatch_queue_t queue, dispatch_block_t block)
 }
 
 /**
- * Safely submits a block to the main dispatch queue for asynchronous execution
- * and returns immediately.
- */
-NS_INLINE void es_dispatch_async_main(dispatch_block_t block)
-{
-    es_dispatch_async(dispatch_get_main_queue(), block);
-}
-
-/**
- * Safely submits a block to the main dispatch queue for synchronous execution
- * and waits until that block completes.
- */
-NS_INLINE void es_dispatch_sync_main(DISPATCH_NOESCAPE dispatch_block_t block)
-{
-    es_dispatch_sync(dispatch_get_main_queue(), block);
-}
-
-/**
  * Submits a block for asynchronous execution on a system-defined global
  * concurrent queue with the specified quality of service.
  */