+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## master
 
 - Added helpers: `es_dispatch_is_queue()`, `es_dispatch_async()`, `es_dispatch_sync()`
+- Removed `es_dispatch_async_main()`, use `es_dispatch_async(dispatch_get_main_queue(), block)` instead
+- Removed `es_dispatch_sync_main()`, use `es_dispatch_sync(dispatch_get_main_queue(), block)` instead
 
 ## 3.17.3 (2019-08-02)